requirements.txt
# ── Core Framework ──────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
pydantic==2.10.4
pydantic-settings==2.7.1

# ── Database ────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1
psycopg2-binary==2.9.10

# ── Cache / State ────────────────────────────────────────────────
redis[hiredis]==5.2.1

# ── Task Queue ──────────────────────────────────────────────────
celery[redis]==5.4.0

# ── ML / RL (numpy-only DQN — no torch needed) ──────────────────
numpy==2.2.1
scipy==1.15.0

# ── Observability ───────────────────────────────────────────────
structlog==24.4.0
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0

# ── Security ────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# ── Utilities ───────────────────────────────────────────────────
httpx==0.28.1
python-dotenv==1.0.1
orjson==3.10.13
